Refactor codebase and change public API
This commit is a large-scale refactoring of the Argy-Bargy codebase. It
occurred as a result of an attempt to use Argy-Bargy for the Jeep
project management tool.

In addition to changes to internal functions, the public API has been
reworked. To reflect this, the previously named functions parse and
parse-with-subcommands have been renamed to parse-args and
parse-args-with-subcommands.